creating basket and user models

+++ api/models/basket.go
@@ -2,6 +2,6 @@ package models
 type Basket struct{
 	Id int `db:"Id" json:"id"`
 	UserId int `db:"UserId" json:"userId"`
-	Amount int `db:"Amount" json:"amount"`
+	Amount float64 `db:"Amount" json:"amount"`
 	BasketItems []BasketItem `json:"basketItems"`
 }
